microservicio_reportes/requirements.txt
Flask
pandas
openpyxl
fpdf
mysql-connector-python
pymongo
